requirements.txt
opencv-python
numpy
pycryptodome
deepface 
reedsolo
scikit-image